requirements.txt
google-cloud-bigquery==2.27.0
pandas==1.3.3
db-dtypes==1.2.0
boto3==1.34.34